# Optimize build speed