--- a/package.json
+++ b/package.json
@@ -119,6 +119,7 @@
     "stylelint": "^13.6.1",
     "stylelint-config-recommended-scss": "^4.2.0",
     "stylelint-scss": "^3.18.0",
+    "terser-webpack-plugin": "^4.1.0",
     "to-string-loader": "^1.1.6",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^4.43.0",