--- a/package.json
+++ b/package.json
@@ -98,6 +98,7 @@
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-react": "^7.20.0",
     "file-loader": "^6.0.0",
+    "hard-source-webpack-plugin": "^0.13.1",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^26.0.1",
     "jest-fetch-mock": "^3.0.3",